moved both apis to same file
